requirements.txt
flask
python-dotenv
FLASK-SQLalchemy
flask-migrate
flask-marshmallow
flask-smorest
flask-jwt
flask_jwt_extended
gunicorn